requirements.txt
--index-url https://download.pytorch.org/whl/cu126

torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0

# PyTorch Geometric wheels (must match torch + cuda)
--find-links https://data.pyg.org/whl/torch-2.8.0+cu126.html

torch-scatter
torch-cluster
torch-geometric

# Utilities (from PyPI)
--extra-index-url https://pypi.org/simple

torchinfo
torchmetrics

# training
mmengine==0.10.2
lightning==2.1.3
einops
# tools
scikit-learn==1.4.0
scipy==1.12.0
tqdm==4.64.0
numba==0.58.1
numpy==1.26.3
mmcv-lite==2.1.0
pandas==2.2.0
dm-tree==0.1.8
opencv-python==4.9.0.80
tabulate==0.9.0
tensorboard==2.15.1
lazy_loader==0.3
umap-learn==0.5.5
protobuf==3.20.0
# visualize
jupyterlab==4.0.11
matplotlib==3.8.2
seaborn==0.13.2
# protein
biopython==1.83
biotite==0.39.0
gemmi==0.6.4
mrcfile==1.5.0
starfile
# formatting
autoflake==2.2.1
pre-commit==3.6.0